+++ src/Providers/QueuedJobsServiceProvider.php
@@ -32,6 +32,7 @@ final class QueuedJobsServiceProvider extends ServiceProvider
 
     public function boot(): void
     {
+        $this->loadMigrationsFrom(__DIR__ . '/../../database/migrations');
         if ($this->app->runningInConsole()) {
 
             $this->publishes(